+++ R/dummy_multiply.R
@@ -1,9 +1,0 @@
-#' Multiply two numbers together
-#'
-#' This is a dummy function and should be removed
-#' @param a,b Two numbers to multiply
-#' @return The product of `a` and `b`
-#' @export
-dummy_multiply = function(a, b) {
-  return(a * b)
-}